requirements.txt
# GUI Framework
PySide6>=6.7.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# HTTP & Networking
requests>=2.31.0
httpx>=0.27.0

# Data & Config
PyYAML>=6.0

# Google AI / Translation
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.1

# Build & Distribution
pyinstaller>=6.0.0